# src/main/java/com/srs/dto/ShiftDTO.java
package com.srs.dto;

public class ShiftDTO {

    private Integer shiftId;


    private String startDate;


    private String endDate;

    public ShiftDTO() {
    }

    public Integer getShiftId() {
        return shiftId;
    }

    public void setShiftId(Integer shiftId) {
        this.shiftId = shiftId;
    }

    public String getStartDate() {
        return startDate;
    }

    public void setStartDate(String startDate) {
        this.startDate = startDate;
    }

    public String getEndDate() {
        return endDate;
    }

    public void setEndDate(String endDate) {
        this.endDate = endDate;
    }
}