# target/classes/srs-openapi.yaml
openapi: "3.0.3"
info:
  title: "srs API"
  description: "srs API"
  version: "1.0.0"
servers:
  - url: "https://srs"
paths:
...
  /sentences/add:
    post:
      summary: "POST sentences/add"
      operationId: "addSentence"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SentenceDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /sentences/delete/{id}:
    delete:
      summary: "DELETE sentences/delete/{id}"
      operationId: "deleteSentence"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /sentences/filter:
    get:
      summary: "GET sentences/filter"
      operationId: "getSentencesByFilter"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SentenceFilter"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/SentenceDTO"
  /sentences/update:
    put:
      summary: "PUT sentences/update"
      operationId: "updateSentence"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SentenceDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
...
  /shifts/add:
    post:
      summary: "POST shifts/add"
      operationId: "addShift"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /shifts/delete/{id}:
    delete:
      summary: "DELETE shifts/delete/{id}"
      operationId: "deleteShift"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /shifts/filter:
    get:
      summary: "GET shifts/filter"
      operationId: "getShiftsByFilter"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftFilter"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/ShiftDTO"
  /shifts/update:
    put:
      summary: "PUT shifts/update"
      operationId: "updateShift"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
...
components:
  schemas:
...
    OperationStatus:
      type: "object"
      properties:
        targetClassName:
          type: "string"
        recordId:
          type: "integer"
          format: "int32"
        operationName:
          type: "string"
        success:
          type: "boolean"
        msgList:
          type: "array"
          items:
            type: "string"
...
    SentenceDTO:
      type: "object"
      properties:
        sentenceId:
          type: "integer"
          format: "int32"
        prisonerId:
          type: "integer"
          format: "int32"
        prisonId:
          type: "integer"
          format: "int32"
        startDate:
          type: "string"
          format: "date"
        endDate:
          type: "string"
          format: "date"
        assignDate:
          type: "string"
          format: "date"
        unassignDate:
          type: "string"
          format: "date"
    SentenceFilter:
      type: "object"
      properties: { }
    ShiftDTO:
      type: "object"
      properties:
        shiftId:
          type: "integer"
          format: "int32"
        startDate:
          type: "string"
          format: "date-time"
        endDate:
          type: "string"
          format: "date-time"
    ShiftFilter:
      type: "object"
      properties: { }
...